Move MockNet to ya-framework-mocks
