+++ core/market/Cargo.toml
@@ -73,5 +73,7 @@ serial_test = { git = "https://github.com/golemfactory/serial_test.git", branch 
 structopt = "0.3"
 tokio = { version = "1", features = ["macros", "rt"] }
 
+ya-framework-mocks = "0.1"
+
 [lints]
 workspace = true
