+++ test-utils/test-framework/framework-basic/src/mocks.rs
@@ -1,2 +1,1 @@
-pub mod bcast;
 pub mod net;
