+++ test-utils/test-framework/framework-basic/src/mocks/net.rs
@@ -1,238 +1,23 @@
 use anyhow::Result;
-use std::collections::HashMap;
-use std::rc::Rc;
-use std::sync::{Arc, Mutex};
+use std::sync::Arc;
 
-use ya_client::model::NodeId;
+use ya_client_model::NodeId;
 use ya_core_model::net;
-use ya_core_model::net::{local as local_net, local::SendBroadcastMessage};
-use ya_net::hybrid::testing::{parse_from_to_addr, parse_net_to_addr};
-use ya_service_bus::{serialization, typed as bus, untyped as local_bus, Error, RpcMessage};
 
-#[cfg(feature = "bcast-singleton")]
-use super::bcast::singleton::BCastService;
-use super::bcast::BCast;
-#[cfg(not(feature = "bcast-singleton"))]
-use super::bcast::BCastService;
-use ya_core_model::net::local::SendBroadcastStub;
-
-#[derive(Clone)]
-pub struct MockNet {
-    inner: Arc<Mutex<MockNetInner>>,
+pub trait IMockBroadcast {
+    /// Registers node to be visible only in specific subnet
+    fn register_for_broadcasts(&self, _node_id: &NodeId, _subnet: &str);
+    /// Subscribes endpoint to topic; endpoint prefix is /<subnet>/
+    fn subscribe_topic(&self, subscribe: net::local::Subscribe);
+    /// Returns all nodes with same subnet as given node subscribed to given topic
+    fn resolve(&self, node_id: &str, topic: &str) -> Vec<Arc<str>>;
 }
 
-#[derive(Default)]
-struct MockNetInner {
-    /// Maps NodeIds to gsb prefixes of other nodes.
-    pub nodes: HashMap<NodeId, String>,
+pub trait IMockNet: IMockBroadcast {
+    fn bind_gsb(&self);
+    fn register_node(&self, node_id: &NodeId, prefix: &str);
+    fn unregister_node(&self, node_id: &NodeId) -> Result<()>;
 }
-
-lazy_static::lazy_static! {
-    static ref NET : MockNet = MockNet {
-        inner: Arc::new(Mutex::new(MockNetInner::default()))
-    };
-}
-
-impl Default for MockNet {
-    fn default() -> Self {
-        log::debug!("Getting singleton MockNet");
-        (*NET).clone()
-    }
-}
-
-impl MockNet {
-    pub fn bind_gsb(&self) {
-        let inner = self.inner.lock().unwrap();
-        inner.bind_gsb()
-    }
-
-    pub fn register_node(&self, node_id: &NodeId, prefix: &str) {
-        log::info!("[MockNet] Registering node {node_id} at prefix: {prefix}");
-
-        let mut inner = self.inner.lock().unwrap();
-        if inner.nodes.insert(*node_id, prefix.to_string()).is_some() {
-            panic!("[MockNet] Node [{}] already existed.", node_id);
-        }
-    }
-
-    pub fn unregister_node(&self, node_id: &NodeId) -> Result<()> {
-        let mut inner = self.inner.lock().unwrap();
-        inner
-            .nodes
-            .remove(node_id)
-            .map(|_| ())
-            .ok_or_else(|| anyhow::anyhow!("node not registered: {}", node_id))
-    }
-
-    fn translate_to(&self, id: NodeId, addr: &str) -> Result<String> {
-        let prefix = self.node_prefix(id)?;
-        let net_prefix = format!("/net/{}", id);
-        Ok(addr.replacen(&net_prefix, &prefix, 1))
-    }
-
-    fn node_prefix(&self, id: NodeId) -> Result<String> {
-        let inner = self.inner.lock().unwrap();
-        inner
-            .nodes
-            .get(&id)
-            .cloned()
-            .ok_or_else(|| anyhow::anyhow!("Node not registered: {id}"))
-    }
-
-    pub fn node_by_prefix(&self, address: &str) -> Option<NodeId> {
-        let inner = self.inner.lock().unwrap();
-        for (id, prefix) in inner.nodes.iter() {
-            if address.contains(prefix) {
-                return Some(*id);
-            }
-        }
-        None
-    }
-}
-
-// TODO: all tests using this mock net implementation should be run sequentially
-// because GSB router is a static singleton (shared state) and consecutive bindings
-// for same addr (ie. local_net::BUS_ID) are being overwritten and only last is effective
-// which means there might be interlace in BCastService instances being used
-// `bcast::singleton` is a try to handle it, but unsuccessful yet
-impl MockNetInner {
-    pub fn bind_gsb(&self) {
-        let bcast = BCastService::default();
-        log::info!("initializing BCast on mock net");
-
-        let bcast_service_id = <SendBroadcastMessage<()> as RpcMessage>::ID;
-
-        let bcast1 = bcast.clone();
-        let _ = bus::bind(local_net::BUS_ID, move |subscribe: local_net::Subscribe| {
-            let bcast = bcast1.clone();
-            async move {
-                log::debug!("subscribing BCast: {:?}", subscribe);
-                bcast.add(subscribe);
-                Ok(0) // ignored id
-            }
-        });
-
-        let addr = format!("{}/{}", local_net::BUS_ID, bcast_service_id);
-        let resp: Rc<[u8]> = serialization::to_vec(&Ok::<(), ()>(())).unwrap().into();
-        let _ = local_bus::subscribe(
-            &addr,
-            move |caller: &str, _addr: &str, msg: &[u8]| {
-                let mock_net = MockNet::default();
-                let resp = resp.clone();
-                let bcast = bcast.clone();
-
-                let stub: SendBroadcastStub = serialization::from_slice(msg).unwrap();
-                let caller = caller.to_string();
-
-                let msg = msg.to_vec();
-
-                let topic = stub.topic;
-                let endpoints = bcast.resolve(&caller, &topic);
-
-                log::debug!("BCasting on {topic} to {endpoints:?} from {caller}");
-                for endpoint in endpoints {
-                    let addr = format!("{endpoint}/{bcast_service_id}");
-
-                    // Normal net would have additional step: Broadcast message would be sent to other node first on /net/{node_id}.
-                    // Net would receive message, check topic and translate it to local addresses interested in this topic.
-                    // Here for simplicity we are skipping those additional steps and directly sending to all endpoints waiting for broadcast.
-                    //
-                    // But since all broadcast handlers are bound on `/local` and all addresses registered in net are on `/public`,
-                    // we must replace `local` -> `public` to find NodeId of receiver.
-                    let addr_local = addr.replacen("local", "public", 1);
-
-                    let node_id = match mock_net.node_by_prefix(&addr_local) {
-                        Some(node_id) => node_id,
-                        None => {
-                            log::debug!(
-                                "Not broadcasting on topic {topic} to {addr}. Node not found on list. \
-                                Probably networking was disabled for this Node."
-                            );
-                            continue;
-                        }
-                    };
-
-                    log::debug!("BCasting on {topic} to address: {addr}, node: [{node_id}]");
-
-                    let caller = caller.clone();
-                    let msg = msg.clone();
-                    tokio::task::spawn_local(async move {
-                        let _ = local_bus::send(addr.as_ref(), &caller, msg.as_ref()).await;
-                    });
-                }
-                async move { Ok(Vec::from(resp.as_ref())) }
-            },
-            (),
-        );
-
-        Self::bind_local_bus(MockNet::default(), FROM_BUS_ID, from_address_resolver);
-        Self::bind_local_bus(MockNet::default(), FROM_UDP_BUS_ID, from_address_resolver);
-        Self::bind_local_bus(
-            MockNet::default(),
-            FROM_TRANSFER_BUS_ID,
-            from_address_resolver,
-        );
-
-        Self::bind_local_bus(MockNet::default(), net::BUS_ID, net_address_resolver);
-        Self::bind_local_bus(MockNet::default(), net::BUS_ID_UDP, net_address_resolver);
-        Self::bind_local_bus(
-            MockNet::default(),
-            net::BUS_ID_TRANSFER,
-            net_address_resolver,
-        );
-    }
-
-    fn bind_local_bus<F>(net: MockNet, address: &'static str, resolver: F)
-    where
-        F: Fn(&str, &str) -> anyhow::Result<(String, NodeId, String)> + 'static,
-    {
-        let resolver = Arc::new(resolver);
-
-        local_bus::subscribe(
-            address,
-            move |caller: &str, addr: &str, msg: &[u8]| {
-                let mock_net = net.clone();
-                let data = Vec::from(msg);
-                let caller = caller.to_string();
-                let addr = addr.to_string();
-                let resolver_ = resolver.clone();
-
-                async move {
-                    log::info!("[MockNet] Received message from [{caller}], on address [{addr}].");
-
-                    let (from, to, address) = resolver_(&caller, &addr)
-                        .map_err(|e| Error::GsbBadRequest(e.to_string()))?;
-                    let translated = mock_net
-                        .translate_to(to, &address)
-                        .map_err(|e| Error::GsbBadRequest(e.to_string()))?;
-
-                    log::info!(
-                        "[MockNet] Sending message from [{from}], to: [{to}], address [{translated}]."
-                    );
-                    local_bus::send(&translated, &from.to_string(), &data).await
-                }
-            },
-            // TODO: Implement stream handler
-            (),
-        );
-    }
-}
-
-fn from_address_resolver(_caller: &str, addr: &str) -> anyhow::Result<(String, NodeId, String)> {
-    let (from, to, addr) =
-        parse_from_to_addr(addr).map_err(|e| anyhow::anyhow!("invalid address: {}", e))?;
-    Ok((from.to_string(), to, addr))
-}
-
-fn net_address_resolver(caller: &str, addr: &str) -> anyhow::Result<(String, NodeId, String)> {
-    let (to, addr) =
-        parse_net_to_addr(addr).map_err(|e| anyhow::anyhow!("invalid address: {}", e))?;
-    Ok((caller.to_string(), to, addr))
-}
-
-pub(crate) const FROM_BUS_ID: &str = "/from";
-pub(crate) const FROM_UDP_BUS_ID: &str = "/udp/from";
-pub(crate) const FROM_TRANSFER_BUS_ID: &str = "/transfer/from";
 
 pub fn gsb_prefixes(test_name: &str, name: &str) -> (String, String) {
     let public_gsb_prefix = format!("/{}/{}/public", test_name, name);
