+++ test-utils/test-framework/framework-mocks/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod identity;
 pub mod market;
+pub mod net;
 pub mod node;
 pub mod payment;
